Implement authorization and validation for Zone management; add ZoneP…

+++ projecteBackend/app/Http/Controllers/ZoneController.php
@@ -4,11 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreZoneRequest;
 use App\Http\Requests\UpdateZoneRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Models\Zone;
 use Illuminate\Http\Request;
 
 class ZoneController extends Controller
 {
+
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -23,6 +27,7 @@ class ZoneController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Zone::class);
         return view('webzones.create');
     }
 
@@ -31,7 +36,13 @@ class ZoneController extends Controller
      */
     public function store(StoreZoneRequest $request)
     {
-        //
+        $this->authorize('create', Zone::class);
+        $validated = $request->validated();
+
+        
+        Zone::create($validated);
+
+        return redirect()->route('webzones.index')->with('success', 'Zona creada correctament!');
     }
 
     /**
@@ -47,6 +58,7 @@ class ZoneController extends Controller
      */
     public function edit(Zone $zone)
     {
+        $this->authorize('update', $zone);
         return view('webzones.edit', compact('zone'));
     }
 
@@ -55,10 +67,11 @@ class ZoneController extends Controller
      */
     public function update(UpdateZoneRequest $request, Zone $zone)
     {
+        $this->authorize('update', $zone);
         $validated = $request->validated();
         $equip = Zone::findOrFail($zone->id);
         $equip->update($validated);
-        return redirect()->route('webzones.index')->with('success', 'Equip actualitzat correctament!');
+        return redirect()->route('webzones.index')->with('success', 'Zona actualitzada correctament!');
     }
 
     /**
@@ -66,6 +79,7 @@ class ZoneController extends Controller
      */
     public function destroy(Zone $zone)
     {
+        $this->authorize( 'delete', $zone);
         $zone->delete();
         return redirect()->route('webzones.index')->with('success', 'Zona esborrada correctament!');
     }
